spic: avoid usage of incremental hash interface
We currently require the user to supply an incremental hash interface,
but we only use this while walking the inclusion proof. With a little
additional copying, we can simplify this to just a monolithinc sha256()
call, simplifying state handling and reducing our API surface.

This is less efficient - with the previous approach, the
bare_read_exact() call could have been replaced by bare_fetch_exact()
and we need one final memmove() to make space for the checkpoint
construction. However, on balance, I think this pales in comparison to
the cryptographic operations we do and the simplification is worth it.